feat: initial implementation
changes:
- refactoring and renaming
