+++ BB84.EntityFrameworkCore.sln
@@ -7,7 +7,7 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{AD1CC5E3-EAB
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BB84.EntityFrameworkCore.Models", "src\BB84.EntityFrameworkCore.Models\BB84.EntityFrameworkCore.Models.csproj", "{EAEB37EA-5EBC-48FE-B113-3A1E3577BE2B}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BB84.EntityFrameworkCore.Repository", "src\BB84.EntityFrameworkCore.Repository\BB84.EntityFrameworkCore.Repository.csproj", "{056520FA-D0E5-4741-9B2C-41BD52AF5F5C}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BB84.EntityFrameworkCore.Repositories", "src\BB84.EntityFrameworkCore.Repositories\BB84.EntityFrameworkCore.Repositories.csproj", "{056520FA-D0E5-4741-9B2C-41BD52AF5F5C}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "tests", "tests", "{C3F6ECA2-55CD-4DA1-BEF7-F9B56ADB84E2}"
 EndProject
@@ -26,9 +26,9 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 		README.md = README.md
 	EndProjectSection
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BB84.EntityFrameworkCore.ModelsTests", "tests\BB84.EntityFrameworkCore.ModelsTests\BB84.EntityFrameworkCore.ModelsTests.csproj", "{86A294B4-8CA4-4C26-833E-55DE7F6090F3}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BB84.EntityFrameworkCore.ModelsTests", "tests\BB84.EntityFrameworkCore.ModelsTests\BB84.EntityFrameworkCore.ModelsTests.csproj", "{86A294B4-8CA4-4C26-833E-55DE7F6090F3}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BB84.EntityFrameworkCore.RepositoryTests", "tests\BB84.EntityFrameworkCore.RepositoryTests\BB84.EntityFrameworkCore.RepositoryTests.csproj", "{B6AE6D2F-6F73-477E-BC83-A408DCD38FA5}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BB84.EntityFrameworkCore.RepositoriesTests", "tests\BB84.EntityFrameworkCore.RepositoriesTests\BB84.EntityFrameworkCore.RepositoriesTests.csproj", "{B6AE6D2F-6F73-477E-BC83-A408DCD38FA5}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
